refactor: 재사용 가능한 메소드 분리 및 메소드 명 변경
- isEmpty() 메소드 StringUtils 클래스 로 이동
- toNumbers() 메소드 ColltionsUtils 클래스 로 이동
- lessThenZero() 메소드 명 isLessThenZero()로 변경
- isLessThenZero() 메소드 NumberUtils 클래스 로 이동
- calculator package 추가
