+++ src/main/java/calculator/ColletionUtils.java
@@ -1,0 +1,12 @@
+package calculator;
+
+import java.util.Arrays;
+
+public class ColletionUtils {
+
+    public static int[] toNumbers(final String[] stringNumbers) {
+        return Arrays.stream(stringNumbers)
+                .mapToInt(Integer::parseInt)
+                .toArray();
+    }
+}
