+++ src/main/java/calculator/NumberUtils.java
@@ -1,0 +1,9 @@
+package calculator;
+
+public class NumberUtils {
+    private static final int ZERO = 0;
+
+    public static <T extends Number> boolean isLessThenZero(final T number) {
+        return number.intValue() < ZERO;
+    }
+}
