+++ src/main/java/calculator/StringCalculator.java
@@ -1,7 +1,9 @@
-package study;
+package calculator;
+
+import static calculator.StringUtils.*;
+import static calculator.ColletionUtils.*;
 
 import java.util.Arrays;
-import java.util.Objects;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
@@ -12,7 +14,7 @@ public class StringCalculator {
     private static final int CUSTOM_DELIMITER = 1;
     private static final int SPLIT_WORD = 2;
     private static final int ZERO = 0;
-    private static final String EMPTY = "";
+
     private static final String NEGATIVE_NUMBER_ERROR = "음수는 사용 불가능 합니다.";
 
     public static int splitAndSum(final String word) {
@@ -23,8 +25,9 @@ public class StringCalculator {
         return sum(toNumbers(split(word)));
     }
 
-    private static boolean isEmpty(final String word) {
-        return word == null || EMPTY.equals(word);
+    private static int sum(final int[] numbers) {
+        validate(numbers);
+        return Arrays.stream(numbers).sum();
     }
 
     private static String[] split(final String word) {
@@ -35,17 +38,6 @@ public class StringCalculator {
         return splitDefault(word);
     }
 
-    private static int sum(final int[] numbers) {
-        isValid(numbers);
-        return Arrays.stream(numbers).sum();
-    }
-
-    private static int[] toNumbers(final String[] stringNumbers) {
-        return Arrays.stream(stringNumbers)
-                .mapToInt(Integer::parseInt)
-                .toArray();
-    }
-
     private static String[] splitCustom(final Matcher matcher) {
         return matcher.group(SPLIT_WORD).split(matcher.group(CUSTOM_DELIMITER));
     }
@@ -54,16 +46,12 @@ public class StringCalculator {
         return word.split(DEFAULT_DELIMITER);
     }
 
-    private static void isValid(final int[] numbers) {
+    private static void validate(final int[] numbers) {
         boolean negativeNumber = Arrays.stream(numbers)
-                .anyMatch(StringCalculator::lessThenZero);
+                .anyMatch(NumberUtils::isLessThenZero);
 
         if (negativeNumber) {
             throw new IllegalArgumentException(NEGATIVE_NUMBER_ERROR);
         }
-    }
-
-    private static <T extends Number> boolean lessThenZero(final T number) {
-        return number.intValue() < ZERO;
     }
 }
