+++ src/test/java/calculator/StringCalculatorTest.java
@@ -1,5 +1,6 @@
-package study;
+package calculator;
 
+import calculator.StringCalculator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
